Moved HashInnerJoinBatchIterator to the sql module
