+++ sql/src/main/java/io/crate/execution/engine/join/HashInnerJoinBatchIterator.java
@@ -20,13 +20,15 @@
  * agreement.
  */
 
-package io.crate.data.join;
+package io.crate.execution.engine.join;
 
 import com.carrotsearch.hppc.IntObjectHashMap;
 import io.crate.data.BatchIterator;
 import io.crate.data.Paging;
 import io.crate.data.Row;
 import io.crate.data.UnsafeArrayRow;
+import io.crate.data.join.ElementCombiner;
+import io.crate.data.join.JoinBatchIterator;
 
 import java.util.ArrayList;
 import java.util.Iterator;
